feat: add dual CLI mode support (#14)
This PR add supports for dual CLI mode.
In normal case, the CLI will expose the ota-image-builder interface. But when the single binary CLI is called with name ota-image-tools.*, it will reveal the CLI interface of ota-image-tools.

A common way to use dual CLI mode is by creating ota-image-tools symlink points to the ota-image-builder bin.